Remove resetting the sample rate for the fixed timer
Resetting the timer frequency takes around 3us.
This is unnecessary for the fixed timer.
We just need to tell the users to keep the same sample
rate for every additional sound (which we kind of already do.)

Else you would have issues anyways, when playing e.g. a horn
and indicator light ticks at the same time.