Fix sound volume calculation
Three issues are fixed:

1.
in a lot of cases the last sample is not handled.
if (sampleCount < sampleLength-1)
will omit the last sample.

2.
Multiplying the volumes at the end for samples
that are not even outputted makes no sense.
Actually, since all the sounds have a sound volume
it is unnecessary.

3.
Adding the samples directly to a buffer is faster
than first adding it to variables.